backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-anymail==12.0
django-ckeditor==6.7.3
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-shortuuidfield==0.1.3
django-smart-selects==1.7.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
environs==14.1.1
idna==3.10
inflection==0.5.1
lxml==5.4.0
marshmallow==3.26.1
packaging==24.2
pillow==11.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
shortuuid==1.0.13
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
